fix(ARC-2496): change metadata order and rendering + split object detail
https://meemoo.atlassian.net/browse/ARC-2496
