+++ src/modules/admin/views/AdminMaterialRequests.tsx
@@ -155,9 +155,9 @@ export const AdminMaterialRequests: FC<DefaultSeoInfo> = ({ url }) => {
 	const renderPagination = ({ gotoPage }: { gotoPage: (i: number) => void }): ReactNode => (
 		<PaginationBar
 			{...getDefaultPaginationBarProps()}
-			start={Math.max(0, filters.page - 1) * ADMIN_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
-			total={materialRequests?.total || 0}
-			count={ADMIN_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
+			startItem={Math.max(0, filters.page - 1) * ADMIN_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
+			totalItems={materialRequests?.total || 0}
+			itemsPerPage={ADMIN_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
 			onPageChange={(pageZeroBased: number) => onPageChange(pageZeroBased, gotoPage)}
 		/>
 	);
