+++ src/modules/cp/views/CpAdminMaterialRequests.tsx
@@ -143,9 +143,9 @@ export const CpAdminMaterialRequests: FC<DefaultSeoInfo> = ({ url }) => {
 	const renderPagination = ({ gotoPage }: { gotoPage: (i: number) => void }): ReactNode => (
 		<PaginationBar
 			{...getDefaultPaginationBarProps()}
-			start={Math.max(0, filters.page - 1) * CP_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
-			total={materialRequests?.total || 0}
-			count={CP_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
+			startItem={Math.max(0, filters.page - 1) * CP_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
+			totalItems={materialRequests?.total || 0}
+			itemsPerPage={CP_MATERIAL_REQUESTS_TABLE_PAGE_SIZE}
 			onPageChange={(pageZeroBased: number) => onPageChange(pageZeroBased, gotoPage)}
 		/>
 	);
