+++ src/modules/cp/views/CpAdminVisitors.tsx
@@ -200,9 +200,9 @@ export const CpAdminVisitorsPage: FC<DefaultSeoInfo> = ({ url }) => {
 						return (
 							<PaginationBar
 								{...getDefaultPaginationBarProps()}
-								count={RequestTablePageSize}
-								start={Math.max(0, filters.page - 1) * RequestTablePageSize}
-								total={visits?.total || 0}
+								itemsPerPage={RequestTablePageSize}
+								startItem={Math.max(0, filters.page - 1) * RequestTablePageSize}
+								totalItems={visits?.total || 0}
 								onPageChange={(pageZeroBased) => {
 									gotoPage(pageZeroBased);
 									// setSelected(null);
