+++ src/modules/ie-objects/ObjectDetailPage.tsx
@@ -1,32 +1,14 @@
 import {
 	Alert,
-	type Breadcrumb,
-	Breadcrumbs,
 	Button,
-	Dropdown,
-	DropdownButton,
-	DropdownContent,
 	FlowPlayer,
 	type FlowPlayerProps,
-	MenuContent,
 	type TabProps,
 	Tabs,
-	TagList,
 } from '@meemoo/react-components';
 import clsx from 'clsx';
 import { type HTTPError } from 'ky';
-import {
-	capitalize,
-	compact,
-	indexOf,
-	intersection,
-	isEmpty,
-	isNil,
-	isString,
-	lowerCase,
-	noop,
-	sortBy,
-} from 'lodash-es';
+import { capitalize, compact, intersection, isNil, lowerCase, noop } from 'lodash-es';
 import getConfig from 'next/config';
 import Head from 'next/head';
 import Link from 'next/link';
@@ -54,22 +36,18 @@ import {
 
 import { GroupName, Permission } from '@account/const';
 import { selectUser } from '@auth/store/user';
+import { type User } from '@auth/types';
 import {
 	RequestAccessBlade,
 	type RequestAccessFormState,
 } from '@home/components/RequestAccessBlade';
 import { useCreateVisitRequest } from '@home/hooks/create-visit-request';
 import { CollapsableBlade } from '@ie-objects/components/CollapsableBlade';
-import { CopyrightConfirmationModal } from '@ie-objects/components/CopyrightConfirmationModal';
-import {
-	type ActionItem,
-	DynamicActionMenu,
-	type DynamicActionMenuProps,
-} from '@ie-objects/components/DynamicActionMenu';
 import { FragmentSlider } from '@ie-objects/components/FragmentSlider';
-import { Metadata, type MetadataItem } from '@ie-objects/components/Metadata';
 import MetadataList from '@ie-objects/components/Metadata/MetadataList';
-import { NamesList } from '@ie-objects/components/NamesList/NamesList';
+import Metadata, {
+	ObjectDetailPageMetadata,
+} from '@ie-objects/components/ObjectDetailPageMetadata/ObjectDetailPageMetadata';
 import { ObjectPlaceholder } from '@ie-objects/components/ObjectPlaceholder';
 import { type MediaObject, RelatedObject } from '@ie-objects/components/RelatedObject';
 import { useGetAltoJsonFileContent } from '@ie-objects/hooks/get-alto-json-file-content';
@@ -78,26 +56,16 @@ import { useGetIeObjectsRelated } from '@ie-objects/hooks/get-ie-objects-related
 import { useGetIeObjectsAlsoInteresting } from '@ie-objects/hooks/get-ie-objects-similar';
 import { useGetIeObjectsTicketInfo } from '@ie-objects/hooks/get-ie-objects-ticket-url';
 import {
-	ANONYMOUS_ACTION_SORT_MAP,
-	CP_ADMIN_ACTION_SORT_MAP,
 	FLOWPLAYER_AUDIO_FORMATS,
 	FLOWPLAYER_FORMATS,
 	FLOWPLAYER_VIDEO_FORMATS,
-	GET_METADATA_FIELDS,
-	GET_NEWSPAPER_DOWNLOAD_OPTIONS,
 	IMAGE_API_FORMATS,
 	IMAGE_FORMATS,
 	JSON_FORMATS,
-	KEY_USER_ACTION_SORT_MAP,
-	KIOSK_ACTION_SORT_MAP,
-	MEDIA_ACTIONS,
-	MEEMOO_ADMIN_ACTION_SORT_MAP,
-	METADATA_EXPORT_OPTIONS,
 	noLicensePlaceholder,
 	OBJECT_DETAIL_TABS,
 	objectPlaceholder,
 	ticketErrorPlaceholder,
-	VISITOR_ACTION_SORT_MAP,
 	XML_FORMATS,
 } from '@ie-objects/ie-objects.consts';
 import {
@@ -108,44 +76,29 @@ import {
 	IeObjectLicense,
 	type IeObjectPageRepresentation,
 	type IeObjectRepresentation,
-	IsPartOfKey,
 	MediaActions,
-	MetadataExportFormats,
-	type MetadataSortMap,
 	ObjectDetailTabs,
 	type OcrSearchResult,
 	type RelatedIeObject,
 } from '@ie-objects/ie-objects.types';
 import {
-	IE_OBJECTS_SERVICE_BASE_URL,
 	IE_OBJECTS_SERVICE_EXPORT,
 	NEWSPAPERS_SERVICE_BASE_URL,
 } from '@ie-objects/services/ie-objects/ie-objects.service.const';
-import { isInAFolder } from '@ie-objects/utils/folders';
-import { getIeObjectRightsOwnerAsText } from '@ie-objects/utils/get-ie-object-rights-owner-as-text';
-import { getIeObjectRightsStatusInfo } from '@ie-objects/utils/get-ie-object-rights-status';
-import { mapKeywordsToTags, renderKeywordsAsTags } from '@ie-objects/utils/map-metadata';
+import { getExternalMaterialRequestUrlIfAvailable } from '@ie-objects/utils/get-external-form-url';
 import IiifViewer from '@iiif-viewer/IiifViewer';
 import { type IiifViewerFunctions, type ImageInfo, type Rect } from '@iiif-viewer/IiifViewer.types';
 import { SearchInputWithResultsPagination } from '@iiif-viewer/components/SearchInputWithResults/SearchInputWithResultsPagination';
 import { MaterialRequestsService } from '@material-requests/services';
-import { useGetAccessibleVisitorSpaces } from '@navigation/components/Navigation/hooks/get-accessible-visitor-spaces';
-import { Blade } from '@shared/components/Blade/Blade';
-import Callout from '@shared/components/Callout/Callout';
 import { ErrorNoAccessToObject } from '@shared/components/ErrorNoAccessToObject';
 import { ErrorNotFound } from '@shared/components/ErrorNotFound';
 import { ErrorSpaceNoLongerActive } from '@shared/components/ErrorSpaceNoLongerActive';
-import HighlightSearchTerms from '@shared/components/HighlightedMetadata/HighlightSearchTerms';
-import HighlightedMetadata from '@shared/components/HighlightedMetadata/HighlightedMetadata';
 import { Icon } from '@shared/components/Icon';
 import { IconNamesLight } from '@shared/components/Icon/Icon.enums';
 import { Loading } from '@shared/components/Loading';
-import MetaDataFieldWithHighlightingAndMaxLength from '@shared/components/MetaDataFieldWithHighlightingAndMaxLength/MetaDataFieldWithHighlightingAndMaxLength';
-import NextLinkWrapper from '@shared/components/NextLinkWrapper/NextLinkWrapper';
-import { Pill } from '@shared/components/Pill';
 import { RedFormWarning } from '@shared/components/RedFormWarning/RedFormWarning';
 import { SeoTags } from '@shared/components/SeoTags/SeoTags';
-import { KNOWN_STATIC_ROUTES, ROUTES_BY_LOCALE } from '@shared/const';
+import { ROUTES_BY_LOCALE } from '@shared/const';
 import {
 	HIGHLIGHTED_SEARCH_TERMS_SEPARATOR,
 	QUERY_PARAM_KEY,
@@ -153,8 +106,7 @@ import {
 import { convertDurationStringToSeconds } from '@shared/helpers/convert-duration-string-to-seconds';
 import { tHtml, tText } from '@shared/helpers/translate';
 import { useHasAnyGroup } from '@shared/hooks/has-group';
-import { useHasAllPermission, useHasAnyPermission } from '@shared/hooks/has-permission';
-import { useIsKeyUser } from '@shared/hooks/is-key-user';
+import { useHasAllPermission } from '@shared/hooks/has-permission';
 import { useGetPeakFile } from '@shared/hooks/use-get-peak-file/use-get-peak-file';
 import { useHideFooter } from '@shared/hooks/use-hide-footer';
 import { useLocale } from '@shared/hooks/use-locale/use-locale';
@@ -162,8 +114,7 @@ import { useStickyLayout } from '@shared/hooks/use-sticky-layout';
 import { useWindowSizeContext } from '@shared/hooks/use-window-size-context';
 import { EventsService, LogEventType } from '@shared/services/events-service';
 import { toastService } from '@shared/services/toast-service';
-import { selectFolders } from '@shared/store/ie-objects';
-import { selectBreadcrumbs, setShowAuthModal, setShowZendesk } from '@shared/store/ui';
+import { setShowAuthModal, setShowZendesk } from '@shared/store/ui';
 import { Breakpoints } from '@shared/types';
 import { IeObjectType } from '@shared/types/ie-objects';
 import { type DefaultSeoInfo } from '@shared/types/seo';
@@ -172,17 +123,10 @@ import { useGetActiveVisitRequestForUserAndSpace } from '@visit-requests/hooks/g
 import { VisitorLayout } from '@visitor-layout/index';
 import { AddToFolderBlade } from '@visitor-space/components/AddToFolderBlade';
 import { MaterialRequestBlade } from '@visitor-space/components/MaterialRequestBlade/MaterialRequestBlade';
-import { NoServerSideRendering } from '@visitor-space/components/NoServerSideRendering/NoServerSideRendering';
 import { VisitorSpaceNavigation } from '@visitor-space/components/VisitorSpaceNavigation/VisitorSpaceNavigation';
 import { ReportBlade } from '@visitor-space/components/reportBlade';
 import { useGetVisitorSpace } from '@visitor-space/hooks/get-visitor-space';
-import {
-	FILTER_LABEL_VALUE_DELIMITER,
-	SearchFilterId,
-	VisitorSpaceStatus,
-} from '@visitor-space/types';
-
-import CopyButton from '../shared/components/CopyButton/CopyButton';
+import { VisitorSpaceStatus } from '@visitor-space/types';
 
 import styles from './ObjectDetailPage.module.scss';
 
@@ -195,25 +139,17 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 	const router = useRouter();
 	const locale = useLocale();
 	const dispatch = useDispatch();
-	const user = useSelector(selectUser);
+	const user: User | null = useSelector(selectUser);
 	const { mutateAsync: createVisitRequest } = useCreateVisitRequest();
-	const breadcrumbs = useSelector(selectBreadcrumbs);
 	const ieObjectId = router.query.ie as string;
 
 	// User types
-	const isKeyUser = useIsKeyUser();
-	const isMeemooAdmin = useHasAnyGroup(GroupName.MEEMOO_ADMIN);
 	const isAnonymous = useHasAnyGroup(GroupName.ANONYMOUS);
-	const isCPAdmin = useHasAnyGroup(GroupName.CP_ADMIN);
 	const isKiosk = useHasAnyGroup(GroupName.KIOSK_VISITOR);
 
 	// Permissions
-	const showResearchWarning = useHasAllPermission(Permission.SHOW_RESEARCH_WARNING);
 	const showLinkedSpaceAsHomepage = useHasAllPermission(Permission.SHOW_LINKED_SPACE_AS_HOMEPAGE);
 	const canManageFolders: boolean | null = useHasAllPermission(Permission.MANAGE_FOLDERS);
-	const canRequestMaterial: boolean | null = useHasAllPermission(
-		Permission.CREATE_MATERIAL_REQUESTS
-	);
 
 	// Internal state
 	const [mediaType, setMediaType] = useState<IeObjectType | null>(null);
@@ -221,11 +157,7 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 	const [hasMediaPlayed, setHasMediaPlayed] = useState(false);
 	const [flowPlayerKey, setFlowPlayerKey] = useState<string | null>(null);
 	const [similar, setSimilar] = useState<MediaObject[]>([]);
-	const [metadataExportDropdownOpen, setMetadataExportDropdownOpen] = useState(false);
-	const [selectedMetadataField, setSelectedMetadataField] = useState<MetadataItem | null>(null);
 	const [isRelatedObjectsBladeOpen, setIsRelatedObjectsBladeOpen] = useState(false);
-	const [copyrightModalOpen, setCopyrightModalOpen] = useState(false);
-	const [onConfirmCopyright, setOnConfirmCopyright] = useState<() => void>(noop);
 	const [hasNewsPaperBeenRendered, setHasNewsPaperBeenRendered] = useState(false);
 	const [hasAppliedUrlSearchTerms, setHasAppliedUrlSearchTerms] = useState<boolean>(false);
 
@@ -235,7 +167,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 
 	// Sizes
 	const windowSize = useWindowSizeContext();
-	const folders = useSelector(selectFolders);
 
 	// Query params
 
@@ -385,12 +316,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 		isLoading: visitorSpaceIsLoading,
 	} = useGetVisitorSpace(router.query.slug as string, false);
 
-	// spaces
-	const canViewAllSpaces = useHasAllPermission(Permission.READ_ALL_SPACES);
-	const { data: accessibleVisitorSpaces } = useGetAccessibleVisitorSpaces({
-		canViewAllSpaces,
-	});
-
 	// ocr alto info
 	const currentPageAltoUrl = useMemo((): string | null => {
 		let altoFileUrl: string | null = null;
@@ -430,23 +355,12 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 			IeObjectAccessThrough.VISITOR_SPACE_FOLDERS,
 			IeObjectAccessThrough.VISITOR_SPACE_FULL,
 		]).length > 0;
-	const isPublicNewspaper: boolean =
-		!!mediaInfo?.licenses?.includes(IeObjectLicense.PUBLIEK_CONTENT) && isNewspaper;
-	const canRequestAccess =
-		isNil(
-			accessibleVisitorSpaces?.find((space) => space.maintainerId === mediaInfo?.maintainerId)
-		) &&
-		mediaInfo?.licenses?.includes(IeObjectLicense.BEZOEKERTOOL_CONTENT) &&
-		isNil(mediaInfo.thumbnailUrl);
-	const showKeyUserPill = mediaInfo?.accessThrough?.includes(IeObjectAccessThrough.SECTOR);
+
 	const showVisitButton =
 		isNil(mediaInfo?.thumbnailUrl) &&
 		mediaInfo?.licenses?.includes(IeObjectLicense.BEZOEKERTOOL_CONTENT) &&
 		visitorSpace?.status === VisitorSpaceStatus.Active &&
 		!isKiosk;
-	const canDownloadMetadata: boolean = useHasAnyPermission(Permission.EXPORT_OBJECT);
-	const canDownloadNewspaper: boolean =
-		useHasAnyPermission(Permission.DOWNLOAD_OBJECT) && isNewspaper && isPublicNewspaper;
 
 	const pageOcrTexts: (string | null)[] = useMemo(() => {
 		const pageOcrTextsTemp: (string | null)[] = [];
@@ -683,13 +597,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 	}, [currentSearchResultIndex]);
 
 	/**
-	 * Close dropdown while resizing
-	 */
-	useEffect(() => {
-		setMetadataExportDropdownOpen(false);
-	}, [windowSize]);
-
-	/**
 	 * Hide the zendesk button for
 	 * - kiosk users
 	 * - users with access to the visitor space of the object
@@ -844,7 +751,7 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 		}
 	};
 
-	const onClickAction = (id: MediaActions) => {
+	const onClickAction = async (id: MediaActions) => {
 		switch (id) {
 			case MediaActions.Bookmark:
 				if (canManageFolders) {
@@ -862,48 +769,10 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 				openRequestAccessBlade();
 				break;
 			case MediaActions.RequestMaterial:
-				onRequestMaterialClick();
+				await onRequestMaterialClick();
 				break;
 		}
 	};
-
-	/**
-	 * If the maintainer of this ie-object has an external form for material requests, we need to construct that url with certain parameters
-	 * This currently is only the case for UGent and VRT
-	 */
-	const getExternalMaterialRequestUrlIfAvailable = useCallback((): string | null => {
-		if (isAnonymous) {
-			return null;
-		}
-
-		if (mediaInfo?.maintainerFormUrl && user) {
-			// Sometimes we want to encode the url params and sometimes we dont. https://meemoo.atlassian.net/browse/ARC-1710
-			// For UGent the whole url is inside one param, so we cannot encode it
-			// For vrt the url is stored in the form_url and the params should be encoded
-			const encodeOrNotUriComponent = mediaInfo.maintainerFormUrl.startsWith('http')
-				? encodeURIComponent
-				: (param: string) => param;
-			return mediaInfo.maintainerFormUrl
-				.replaceAll('{first_name}', encodeOrNotUriComponent(user.firstName))
-				.replaceAll('{last_name}', encodeOrNotUriComponent(user.lastName))
-				.replaceAll('{email}', encodeOrNotUriComponent(user.email))
-				.replaceAll(
-					'{local_cp_id}',
-					encodeOrNotUriComponent(mediaInfo?.meemooLocalId || '')
-				)
-				.replaceAll('{pid}', encodeOrNotUriComponent(mediaInfo?.schemaIdentifier || ''))
-				.replaceAll('{title}', encodeOrNotUriComponent(mediaInfo?.name || ''))
-				.replaceAll(
-					'{title_serie}',
-					encodeOrNotUriComponent(
-						mediaInfo?.isPartOf?.find(
-							(isPartOfEntry) => isPartOfEntry?.collectionType === IsPartOfKey.serie
-						)?.name || ''
-					)
-				);
-		}
-		return null;
-	}, [isAnonymous, mediaInfo, user]);
 
 	const onRequestMaterialClick = async () => {
 		if (isAnonymous) {
@@ -927,7 +796,11 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 			return;
 		}
 
-		const externalFormUrl = getExternalMaterialRequestUrlIfAvailable();
+		const externalFormUrl = getExternalMaterialRequestUrlIfAvailable(
+			mediaInfo,
+			isAnonymous,
+			user
+		);
 		if (externalFormUrl) {
 			EventsService.triggerEvent(LogEventType.ITEM_REQUEST, window.location.href, {
 				fragment_id: mediaInfo?.schemaIdentifier,
@@ -942,23 +815,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 			setActiveBlade(MediaActions.RequestMaterial, 'replaceIn');
 		}
 	};
-
-	const handleOnDownloadEvent = useCallback(() => {
-		const path = window.location.href;
-		const eventData = {
-			type: mediaInfo?.dctermsFormat,
-			fragment_id: mediaInfo?.schemaIdentifier,
-			pid: mediaInfo?.schemaIdentifier,
-			user_group_name: user?.groupName,
-			or_id: mediaInfo?.maintainerId,
-		};
-		EventsService.triggerEvent(LogEventType.DOWNLOAD, path, eventData).then(noop);
-	}, [
-		mediaInfo?.dctermsFormat,
-		mediaInfo?.maintainerId,
-		mediaInfo?.schemaIdentifier,
-		user?.groupName,
-	]);
 
 	const handleOnPlay = () => {
 		setIsMediaPaused(false);
@@ -1077,148 +933,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 		iiifViewerReference.current?.iiifGoToHome();
 	};
 
-	const getSortMapByUserType = useCallback((): MetadataSortMap[] => {
-		if (isNil(user)) {
-			return ANONYMOUS_ACTION_SORT_MAP();
-		}
-
-		if (isKeyUser) {
-			return KEY_USER_ACTION_SORT_MAP(canDownloadMetadata || canDownloadNewspaper);
-		}
-
-		if (isKiosk) {
-			return KIOSK_ACTION_SORT_MAP();
-		}
-
-		if (isMeemooAdmin) {
-			return MEEMOO_ADMIN_ACTION_SORT_MAP(canDownloadMetadata || canDownloadNewspaper);
-		}
-
-		if (isCPAdmin) {
-			return CP_ADMIN_ACTION_SORT_MAP(canDownloadMetadata || canDownloadNewspaper);
-		}
-
-		return VISITOR_ACTION_SORT_MAP(canDownloadMetadata || canDownloadNewspaper);
-	}, [
-		canDownloadMetadata,
-		isKeyUser,
-		isMeemooAdmin,
-		isKiosk,
-		user,
-		isCPAdmin,
-		canDownloadNewspaper,
-	]);
-
-	const onExportClick = useCallback(
-		(format: MetadataExportFormats) => {
-			switch (format) {
-				case MetadataExportFormats.fullNewspaperZip:
-					setCopyrightModalOpen(true);
-					setOnConfirmCopyright(() => () => {
-						window.open(
-							`${publicRuntimeConfig.PROXY_URL}/${NEWSPAPERS_SERVICE_BASE_URL}/${ieObjectId}/${IE_OBJECTS_SERVICE_EXPORT}/zip`
-						);
-						handleOnDownloadEvent();
-					});
-					break;
-
-				case MetadataExportFormats.onePageNewspaperZip:
-					setCopyrightModalOpen(true);
-					setOnConfirmCopyright(() => () => {
-						window.open(
-							`${publicRuntimeConfig.PROXY_URL}/${NEWSPAPERS_SERVICE_BASE_URL}/${ieObjectId}/${IE_OBJECTS_SERVICE_EXPORT}/zip?page=${currentPageIndex}`
-						);
-						handleOnDownloadEvent();
-					});
-					break;
-
-				case MetadataExportFormats.xml:
-				case MetadataExportFormats.csv:
-				default:
-					window.open(
-						`${publicRuntimeConfig.PROXY_URL}/${IE_OBJECTS_SERVICE_BASE_URL}/${ieObjectId}/${IE_OBJECTS_SERVICE_EXPORT}/${format}`
-					);
-			}
-			setMetadataExportDropdownOpen(false);
-		},
-		[currentPageIndex, handleOnDownloadEvent, ieObjectId]
-	);
-
-	const renderExportDropdown = useCallback(
-		(isPrimary: boolean) => {
-			const icon = <Icon name={IconNamesLight.Export} aria-hidden />;
-
-			const buttonLabelDesktop = isPublicNewspaper
-				? tText('modules/ie-objects/object-detail-page___download-deze-krant-desktop')
-				: tText('modules/ie-objects/object-detail-page___export-metadata-desktop');
-			const buttonLabelMobile = isPublicNewspaper
-				? tText('modules/ie-objects/object-detail-page___download-deze-krant-mobile')
-				: tText('modules/ie-objects/object-detail-page___export-metadata-mobile');
-
-			const exportOptions = [];
-
-			if (canDownloadNewspaper) {
-				exportOptions.push(...GET_NEWSPAPER_DOWNLOAD_OPTIONS());
-			}
-
-			if (canDownloadMetadata) {
-				exportOptions.push(...METADATA_EXPORT_OPTIONS());
-			}
-
-			return (
-				<div className={styles['p-object-detail__export']}>
-					<Dropdown
-						isOpen={metadataExportDropdownOpen}
-						onOpen={() => setMetadataExportDropdownOpen(true)}
-						onClose={() => setMetadataExportDropdownOpen(false)}
-					>
-						<DropdownButton>
-							{isPrimary ? (
-								<Button
-									variants={[isPrimary ? 'teal' : 'silver']}
-									className={styles['p-object-detail__export']}
-									iconStart={icon}
-									iconEnd={<Icon name={IconNamesLight.AngleDown} aria-hidden />}
-									aria-label={buttonLabelDesktop}
-									title={buttonLabelDesktop}
-								>
-									<span className="u-text-ellipsis u-display-none u-display-block-md">
-										{buttonLabelDesktop}
-									</span>
-									<span className="u-text-ellipsis u-display-none-md">
-										{buttonLabelMobile}
-									</span>
-								</Button>
-							) : (
-								<Button
-									icon={icon}
-									variants={['silver']}
-									aria-label={buttonLabelDesktop}
-									title={buttonLabelDesktop}
-								/>
-							)}
-						</DropdownButton>
-						<DropdownContent>
-							<MenuContent
-								rootClassName="c-dropdown-menu"
-								className={styles['p-object-detail__export-dropdown']}
-								menuItems={exportOptions}
-								onClick={(id) => onExportClick(id as MetadataExportFormats)}
-							/>
-						</DropdownContent>
-					</Dropdown>
-				</div>
-			);
-		},
-		[
-			isPublicNewspaper,
-			canDownloadNewspaper,
-			canDownloadMetadata,
-			metadataExportDropdownOpen,
-			onExportClick,
-		]
-	);
-
 	/**
 	 * Content
 	 */
@@ -1277,66 +991,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 	}, [isMobile, showLinkedSpaceAsHomepage, visitRequest?.endAt]);
 
 	const showVisitorSpaceNavigationBar = !isNil(accessEndDate) || isKiosk;
-
-	const mediaActions: DynamicActionMenuProps = useMemo(() => {
-		const isMobile = !!(windowSize.width && windowSize.width < Breakpoints.md);
-		const originalActions = MEDIA_ACTIONS({
-			isMobile,
-			canManageFolders: canManageFolders || isAnonymous,
-			isInAFolder: isInAFolder(folders, mediaInfo?.schemaIdentifier),
-			canReport: !isKiosk,
-			canRequestAccess: !!canRequestAccess,
-			canRequestMaterial: isAnonymous || canRequestMaterial,
-			canExport: canDownloadMetadata || canDownloadNewspaper || false,
-			externalFormUrl: getExternalMaterialRequestUrlIfAvailable(),
-		});
-
-		// Sort, filter and tweak actions according to the given sort map
-		const sortMap = getSortMapByUserType();
-		const actions: ActionItem[] = sortBy(originalActions.actions, ({ id }: ActionItem) =>
-			indexOf(
-				sortMap.map((d) => d.id),
-				id
-			)
-		)
-			.map((action: ActionItem) => {
-				const existsInSortMap = !isNil(sortMap.find((d) => d.id === action.id));
-				const isPrimary = sortMap.find((d) => action.id === d.id)?.isPrimary ?? false;
-				const showExport =
-					action.id === MediaActions.Export &&
-					(canDownloadMetadata || canDownloadNewspaper);
-
-				return existsInSortMap
-					? {
-							...action,
-							isPrimary,
-							...(showExport && {
-								customElement: renderExportDropdown(isPrimary),
-							}),
-					  }
-					: null;
-			})
-			.filter(Boolean) as ActionItem[];
-
-		return {
-			...originalActions,
-			actions,
-		};
-	}, [
-		windowSize.width,
-		canManageFolders,
-		isAnonymous,
-		folders,
-		mediaInfo?.schemaIdentifier,
-		isKiosk,
-		canRequestAccess,
-		canRequestMaterial,
-		canDownloadNewspaper,
-		canDownloadMetadata,
-		getExternalMaterialRequestUrlIfAvailable,
-		getSortMapByUserType,
-		renderExportDropdown,
-	]);
 
 	const handleIiifViewerSelection = (rect: Rect) => {
 		window.open(
@@ -1539,78 +1193,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 		</li>
 	);
 
-	const renderBreadcrumbs = (): ReactNode => {
-		const defaultBreadcrumbs: Breadcrumb[] = [
-			...(isKiosk
-				? []
-				: [
-						{
-							label: tText('modules/ie-objects/object-detail-page___home'),
-							to: ROUTES_BY_LOCALE[locale].home,
-						},
-				  ]),
-			{
-				label: tText('modules/ie-objects/object-detail-page___zoeken'),
-				to: ROUTES_BY_LOCALE[locale].search,
-			},
-		];
-
-		const staticBreadcrumbs: Breadcrumb[] = !isEmpty(breadcrumbs)
-			? breadcrumbs
-			: defaultBreadcrumbs;
-
-		const dynamicBreadcrumbs: Breadcrumb[] = !isNil(mediaInfo)
-			? [
-					...(hasAccessToVisitorSpaceOfObject
-						? [
-								{
-									label: mediaInfo?.maintainerName,
-									to: isKiosk
-										? ROUTES_BY_LOCALE[locale].search
-										: `${ROUTES_BY_LOCALE[locale].search}?${SearchFilterId.Maintainer}=${mediaInfo?.maintainerSlug}`,
-								},
-						  ]
-						: []),
-					{
-						label: mediaInfo?.name,
-						to: `${ROUTES_BY_LOCALE[locale].search}/${mediaInfo?.maintainerSlug}/${mediaInfo?.schemaIdentifier}`,
-					},
-			  ]
-			: [];
-
-		return (
-			<Breadcrumbs
-				className="u-mt-32"
-				items={[...staticBreadcrumbs, ...dynamicBreadcrumbs]}
-				icon={<Icon name={IconNamesLight.AngleRight} />}
-				linkComponent={NextLinkWrapper}
-			/>
-		);
-	};
-
-	const renderResearchWarning = (): ReactNode => (
-		<Callout
-			className={clsx(styles['p-object-detail__callout'], 'u-pt-32 u-pb-24')}
-			icon={<Icon name={IconNamesLight.Info} aria-hidden />}
-			text={tHtml(
-				'pages/slug/ie/index___door-gebruik-te-maken-van-deze-applicatie-bevestigt-u-dat-u-het-beschikbare-materiaal-enkel-raadpleegt-voor-wetenschappelijk-of-prive-onderzoek'
-			)}
-			action={
-				<Link
-					passHref
-					href={KNOWN_STATIC_ROUTES.kioskConditions}
-					aria-label={tText('pages/slug/index___meer-info')}
-				>
-					<Button
-						className="u-py-0 u-px-8 u-color-neutral u-font-size-14 u-height-auto"
-						label={tHtml('pages/slug/index___meer-info')}
-						variants={['text', 'underline']}
-					/>
-				</Link>
-			}
-		/>
-	);
-
 	const renderIeObjectCards = (
 		type: 'similar' | 'related',
 		items: MediaObject[],
@@ -1637,314 +1219,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 			}
 		</div>
 	);
-
-	const renderMaintainerMetaTitle = ({
-		maintainerName,
-		maintainerLogo,
-		maintainerId,
-	}: IeObject): ReactNode => (
-		<div className={styles['p-object-detail__metadata-maintainer-title']}>
-			<p className={styles['p-object-detail__metadata-label']}>
-				{tText('modules/ie-objects/const/index___aanbieder')}
-			</p>
-			{!isKiosk && !hasAccessToVisitorSpaceOfObject && (
-				<>
-					<NoServerSideRendering>
-						<div className={styles['p-object-detail__sidebar__content-pill']}>
-							<TagList
-								className="u-pt-12"
-								tags={mapKeywordsToTags([maintainerName])}
-								onTagClicked={async () => {
-									await router.push(
-										stringifyUrl({
-											url: ROUTES_BY_LOCALE[locale].search,
-											query: {
-												[SearchFilterId.Maintainers]: [
-													`${maintainerId}${FILTER_LABEL_VALUE_DELIMITER}${maintainerName}`,
-												],
-											},
-										})
-									);
-								}}
-								variants={['clickable', 'silver', 'medium']}
-							/>
-						</div>
-					</NoServerSideRendering>
-					{maintainerLogo && (
-						<div className={styles['p-object-detail__sidebar__content-logo']}>
-							{/* TODO remove this hack once we fully switched to the new graph.organisations table */}
-							{/* eslint-disable-next-line @next/next/no-img-element */}
-							<img src={maintainerLogo} alt={`Logo ${maintainerName}`} />
-						</div>
-					)}
-				</>
-			)}
-		</div>
-	);
-
-	const renderMetaDataActions = (): ReactNode => (
-		<div className="u-pb-24 p-object-detail__actions">
-			<div className="p-object-detail__primary-actions">
-				<DynamicActionMenu {...mediaActions} onClickAction={onClickAction} />
-			</div>
-		</div>
-	);
-
-	const renderVisitButton = (): ReactNode => (
-		<Button
-			label={tText('modules/ie-objects/components/metadata/metadata___plan-een-bezoek')}
-			variants={['dark', 'sm']}
-			className={styles['p-object-detail__visit-button']}
-			onClick={openRequestAccessBlade}
-		/>
-	);
-
-	const renderMaintainerMetaData = ({
-		maintainerDescription,
-		maintainerSiteUrl,
-		maintainerName,
-	}: IeObject): ReactNode => {
-		if (!isKiosk && !hasAccessToVisitorSpaceOfObject) {
-			return (
-				<div className={styles['p-object-detail__sidebar__content-maintainer-data']}>
-					{maintainerDescription && (
-						<p className={styles['p-object-detail__sidebar__content-description']}>
-							{maintainerDescription}
-						</p>
-					)}
-					{maintainerSiteUrl && (
-						<p className={styles['p-object-detail__sidebar__content-link']}>
-							<a href={maintainerSiteUrl} target="_blank" rel="noopener noreferrer">
-								{maintainerSiteUrl}
-							</a>
-							<Icon className="u-ml-8" name={IconNamesLight.Extern} />
-						</p>
-					)}
-					{showVisitButton && isMobile && renderVisitButton()}
-				</div>
-			);
-		} else {
-			return (
-				<div className={styles['p-object-detail__sidebar__content-maintainer-data']}>
-					{maintainerName}
-					{showVisitButton && isMobile && renderVisitButton()}
-				</div>
-			);
-		}
-	};
-
-	const renderKeyUserPill = (): ReactNode => (
-		<div className="u-mt-24">
-			<Pill
-				isExpanded
-				icon={IconNamesLight.Key}
-				label={tText(
-					'pages/bezoekersruimte/visitor-space-slug/object-id/index___voor-sleutelgebruikers'
-				)}
-				className="u-bg-mustard"
-			/>
-		</div>
-	);
-
-	const renderMetaData = () => {
-		if (isNil(mediaInfo)) {
-			return;
-		}
-
-		const showAlert = !mediaInfo.description;
-		const metaDataFields = GET_METADATA_FIELDS(
-			mediaInfo,
-			getFileByType([...FLOWPLAYER_FORMATS, ...IMAGE_API_FORMATS]),
-			simplifiedAltoInfo || null,
-			locale,
-			publicRuntimeConfig.CLIENT_URL
-		).filter(({ data }: MetadataItem): boolean => !!data);
-		const rightsStatusInfo = isNewspaper
-			? getIeObjectRightsStatusInfo(mediaInfo, locale)
-			: null;
-		let rightsAttributionText: string | null = null;
-		if (
-			isNewspaper &&
-			mediaInfo?.licenses?.includes(IeObjectLicense.PUBLIEK_CONTENT) &&
-			rightsStatusInfo
-		) {
-			rightsAttributionText = compact([
-				getIeObjectRightsOwnerAsText(mediaInfo),
-				mediaInfo.datePublished,
-				mediaInfo.name,
-				mediaInfo.maintainerName,
-				rightsStatusInfo.label,
-				publicRuntimeConfig.CLIENT_URL +
-					ROUTES_BY_LOCALE[locale].permalink.replace(':pid', mediaInfo.schemaIdentifier),
-			]).join(', ');
-		}
-
-		return (
-			<>
-				<div className={styles['p-object-detail__metadata-content']}>
-					{showResearchWarning && renderResearchWarning()}
-					{renderBreadcrumbs()}
-					{showKeyUserPill && renderKeyUserPill()}
-					<h3
-						className={clsx(
-							'u-pb-32',
-							styles['p-object-detail__title'],
-							showKeyUserPill ? 'u-pt-8' : 'u-pt-24'
-						)}
-					>
-						<HighlightSearchTerms toHighlight={mediaInfo?.name} />
-					</h3>
-
-					{renderMetaDataActions()}
-
-					{!!rightsAttributionText && (
-						<>
-							<Alert
-								content={tHtml(
-									'modules/ie-objects/object-detail-page___deze-bronvermelding-is-automatisch-gegenereerd-en-kan-fouten-bevatten-a-href-bronvermelding-fouten-meer-info-a'
-								)}
-							/>
-							<Metadata
-								title={tHtml(
-									'modules/ie-objects/object-detail-page___bronvermelding'
-								)}
-								key="metadata-source-attribution"
-								renderRight={
-									<CopyButton text={rightsAttributionText} variants={['white']} />
-								}
-							>
-								<span>{rightsAttributionText}</span>
-							</Metadata>
-						</>
-					)}
-
-					<MetaDataFieldWithHighlightingAndMaxLength
-						title={tText(
-							'modules/visitor-space/utils/metadata/metadata___beschrijving'
-						)}
-						data={mediaInfo.description}
-						className="u-pb-24 u-line-height-1-4 u-font-size-14"
-						onReadMoreClicked={setSelectedMetadataField}
-					/>
-
-					{showAlert && (
-						<Alert
-							className="c-Alert__margin-bottom"
-							icon={<Icon name={IconNamesLight.Info} />}
-							content={tHtml(
-								'pages/bezoekersruimte/visitor-space-slug/object-id/index___geen-beschrijving'
-							)}
-							title=""
-						/>
-					)}
-				</div>
-
-				<MetadataList allowTwoColumns={true}>
-					<Metadata
-						title={renderMaintainerMetaTitle(mediaInfo)}
-						key={`metadata-maintainer`}
-					>
-						{renderMaintainerMetaData(mediaInfo)}
-					</Metadata>
-					{/* other metadata fields */}
-					{metaDataFields.map((item: MetadataItem, index: number) => {
-						if (isString(item.data)) {
-							return (
-								<Metadata
-									title={item.title}
-									key={`metadata-${index}-${item.title}`}
-								>
-									<MetaDataFieldWithHighlightingAndMaxLength
-										title={item.title}
-										data={item.data}
-										onReadMoreClicked={setSelectedMetadataField}
-									/>
-								</Metadata>
-							);
-						} else {
-							return (
-								<Metadata
-									title={item.title}
-									key={`metadata-${index}-${item.title}`}
-								>
-									{item.data}
-								</Metadata>
-							);
-						}
-					})}
-
-					{!!rightsStatusInfo && (
-						<Metadata
-							title={tHtml('modules/ie-objects/object-detail-page___rechten')}
-							className={styles['p-object-detail__metadata-content__rights-status']}
-							key="metadata-rights-status"
-							renderRight={
-								<a target="_blank" href={rightsStatusInfo.link} rel="noreferrer">
-									<Button
-										variants={['white']}
-										icon={<Icon name={IconNamesLight.Extern} />}
-									/>
-								</a>
-							}
-						>
-							<span className="u-flex u-flex-items-center u-gap-xs">
-								{rightsStatusInfo.icon}
-								{rightsStatusInfo?.label}
-								{rightsStatusInfo?.moreInfo}
-							</span>
-						</Metadata>
-					)}
-				</MetadataList>
-
-				<MetadataList allowTwoColumns={false}>
-					{isNewspaper && !!currentPage?.mentions?.length && (
-						<Metadata
-							title={tText('modules/ie-objects/object-detail-page___namenlijst')}
-							key="metadata-fallen-names-list"
-							renderTitleRight={
-								<div className="u-color-neutral u-font-size-14 u-font-weight-400">
-									{tHtml(
-										'modules/ie-objects/object-detail-page___a-href-namenlijst-gesneuvelden-wat-is-dit-a'
-									)}
-								</div>
-							}
-						>
-							<NamesList
-								mentions={currentPage?.mentions || []}
-								onZoomToLocation={iiifViewerReference.current?.iiifZoomTo}
-							/>
-						</Metadata>
-					)}
-
-					{!!mediaInfo.keywords?.length && (
-						<Metadata
-							title={tHtml(
-								'pages/bezoekersruimte/visitor-space-slug/object-id/index___trefwoorden'
-							)}
-							key="metadata-keywords"
-							className="u-pb-0"
-						>
-							{renderKeywordsAsTags(
-								mediaInfo.keywords,
-								visitRequest ? (router.query.slug as string) : '',
-								locale,
-								router
-							)}
-						</Metadata>
-					)}
-					{!!similar.length && (
-						<Metadata
-							title={tHtml('pages/slug/ie/index___ook-interessant')}
-							key="metadata-keywords"
-							className="u-pb-0"
-						>
-							{renderIeObjectCards('similar', similar)}
-						</Metadata>
-					)}
-				</MetadataList>
-			</>
-		);
-	};
 
 	const renderRelatedObjectsBlade = () => {
 		const mappedRelatedIeObjects = getMappedRelatedIeObjects();
@@ -2247,7 +1521,38 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 							}
 						)}
 					>
-						{activeTab === ObjectDetailTabs.Metadata && renderMetaData()}
+						{/*
+						 * IeObject metadata
+						 */}
+						{activeTab === ObjectDetailTabs.Metadata && (
+							<ObjectDetailPageMetadata
+								onClickAction={onClickAction}
+								mediaInfo={mediaInfo}
+								openRequestAccessBlade={openRequestAccessBlade}
+								visitRequest={visitRequest || null}
+								showVisitButton={showVisitButton || false}
+								hasAccessToVisitorSpaceOfObject={hasAccessToVisitorSpaceOfObject}
+								currentPageIndex={currentPageIndex}
+								currentPage={currentPage || null}
+								activeFile={getFileByType([
+									...FLOWPLAYER_FORMATS,
+									...IMAGE_API_FORMATS,
+								])}
+								simplifiedAltoInfo={simplifiedAltoInfo || null}
+								iiifZoomTo={iiifViewerReference.current?.iiifZoomTo}
+							/>
+						)}
+						{!!similar.length && (
+							<MetadataList allowTwoColumns={false}>
+								<Metadata
+									title={tHtml('pages/slug/ie/index___ook-interessant')}
+									key="metadata-keywords"
+									className="u-pb-0"
+								>
+									{renderIeObjectCards('similar', similar)}
+								</Metadata>
+							</MetadataList>
+						)}
 						{activeTab === ObjectDetailTabs.Ocr && renderOcrContent()}
 					</div>
 					{renderRelatedObjectsBlade()}
@@ -2295,26 +1600,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 				onSubmit={onRequestAccessSubmit}
 				id="object-detail-page__request-access-blade"
 			/>
-			{/* Read more metadata field blade */}
-			<Blade
-				className={clsx(
-					'u-pb-24 u-line-height-1-4 u-font-size-14',
-					styles['p-object-detail__metadata-blade']
-				)}
-				isOpen={!!selectedMetadataField}
-				onClose={() => setSelectedMetadataField(null)}
-				renderTitle={(props: Pick<HTMLElement, 'id' | 'className'>) => (
-					<h2 {...props}>{selectedMetadataField?.title}</h2>
-				)}
-				id="object-detail-page__metadata-field-detail-blade"
-			>
-				<div className="u-px-32 u-pb-32">
-					<HighlightedMetadata
-						title={selectedMetadataField?.title}
-						data={selectedMetadataField?.data}
-					/>
-				</div>
-			</Blade>
 		</>
 	);
 
@@ -2406,14 +1691,6 @@ export const ObjectDetailPage: FC<DefaultSeoInfo> = ({ title, description, image
 					imgUrl={image}
 					translatedPages={[]}
 					relativeUrl={url}
-				/>
-				<CopyrightConfirmationModal
-					isOpen={copyrightModalOpen}
-					onClose={() => setCopyrightModalOpen((prevState) => !prevState)}
-					onConfirm={() => {
-						onConfirmCopyright();
-						setCopyrightModalOpen((prevState) => !prevState);
-					}}
 				/>
 				{renderPageContent()}
 			</VisitorLayout>
