+++ src/modules/ie-objects/components/SearchLinkTag/SearchLinkTag.module.scss
@@ -1,0 +1,5 @@
+.c-metadata-search-link {
+	:global(.c-tag--medium) {
+		height: auto !important;
+	}
+}
