+++ src/modules/ie-objects/components/SearchLinkTag/SearchLinkTag.types.ts
@@ -1,0 +1,4 @@
+export interface SearchLinkTagProps {
+	label: string;
+	link: string;
+}
